build on postinstall using babel
Need to transpile source files using babel on postinstall, otherwise it doesn't work in node due to ES6 modules usage. I'm not sure it's the best way to do this, but my thinking is it's better to transpile on postinstall on target machine rather than on publishing. This way babel will transipile code optimally depending on current node version. Btw, this issue wasn't reproducible when I was testing locally using `npm link`.