use prepublish hook instead of postinstall
After some researching came to decision to use "prepublish" hook to keep build files under npm instead of using "postinstall" hook. The benefit is more predictable install and no need to add babel-cli to package dependincies to be installed alongside with package in order to build it on install.

some knowledge sources:
- https://github.com/avajs/ava/issues/184#issuecomment-155713014
- https://docs.npmjs.com/misc/developers#keeping-files-out-of-your-package